Trajectory

+++ pc/src/main/java/group2/sdp/pc/geom/Line.java
@@ -5,7 +5,7 @@ import java.awt.geom.Line2D;
 /**
  * A 2-dimensional line defined by two vector points.
  */
-public class Line extends Line2D.Double {
+public class Line extends Line2D.Float {
 
 	private static final long serialVersionUID = 1L;
